+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Step: 修复自定义 icon 的样式
 - 修复 Tree 组件 checkbox 点击失效的问题
 - Breadcrumb 增加路由跳转的功能
+- DatePicker 修复使用 Tab 键切换时弹出框未隐藏
 
 ### 1.0.0-rc.6